requirements.txt
absl-py==2.2.2
accelerate==0.30.1
addict==2.4.0
aiofiles==23.2.1
aiosignal==1.3.2
altair==5.3.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
blis==0.7.11
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
decorator==4.4.2
decord==0.6.0
deepspeed==0.12.3
editdistance==0.6.2
einops==0.8.0
einx==0.3.0
encodec==0.1.1
et_xmlfile==2.0.0
fairscale==0.4.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
flash_attn==2.7.4.post1
fonttools==4.57.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.3.2
gradio==4.26.0
gradio_client==0.15.1
grpcio==1.71.0
h11==0.16.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
librosa==0.9.0
llvmlite==0.44.0
lxml==5.4.0
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==2.1.5
matplotlib==3.7.4
mdurl==0.1.2
modelscope_studio @ http://thunlp.oss-cn-qingdao.aliyuncs.com/multi_modal/never_delete/modelscope_studio-0.4.0.9-py3-none-any.whl
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
murmurhash==1.0.12
networkx==3.4.2
ninja==1.11.1.1
nltk==3.8.1
numba==0.61.2
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.5.5.64
openpyxl==3.1.2
orjson==3.10.18
packaging==23.2
pandas==2.2.3
peft==0.14.0
Pillow==10.1.0
platformdirs==4.3.7
pooch==1.8.2
portalocker==3.1.1
preshed==3.0.9
proglog==0.1.11
protobuf==4.25.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==13.9.4
rpds-py==0.24.0
ruff==0.11.8
sacrebleu==2.3.2
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.0
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.11
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
socksio==1.0.0
soundfile==0.12.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.6.0
timm==0.9.10
tokenizers==0.19.1
tomlkit==0.12.0
toolz==1.0.0
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.66.1
transformers==4.44.2
triton==2.2.0
typer==0.9.4
typing_extensions==4.8.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.24.0.post1
vector-quantize-pytorch==1.18.5
vocos==0.1.0
wasabi==1.1.3
weasel==0.3.4
websockets==11.0.3
Werkzeug==3.1.3
